remove duplicated CompositeModel + improve some docs
