+++ src/mercs/composition/__init__.py
@@ -1,3 +1,2 @@
 from .compose import o, x
 from .CompositeModel import CompositeModel
-from .NewCompositeModel import NewCompositeModel
